Integrate plaso with timesketch
This is a bit of a hack to integrate them, because when plaso
output to timesketch it wants to write directly to the database.
Plaso gets intalled into the same container as Timesketch. To
easily transfer data to and from the host filesystem, docker-compose
automatically maps hosts /data/ directory to /data/ in container.
Use plazo tools after attaching to container:

$ docker exec -it docker_timesketch_1 /bin/bash